+++ website/src/pages/index.jsx
@@ -1,77 +1,20 @@
 import React from 'react';
 import useBaseUrl from '@docusaurus/useBaseUrl';
-import styled from 'styled-components';
 import Layout from '@theme/Layout';
 
 import {Home} from '../components';
-
-const HeroImage = styled.img`
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  object-position: center;
-  display: block;
-`;
+import styles from './index.module.css';
 
 const HeroExample = () => {
   const src = useBaseUrl('/images/hero-editable-3d-tiles.jpg');
-  return <HeroImage src={src} alt="deck.gl-community editable 3D tiles over the Grand Canyon" />;
+  return (
+    <img
+      className={styles.heroImage}
+      src={src}
+      alt="deck.gl-community editable 3D tiles over the Grand Canyon"
+    />
+  );
 };
-
-const FeatureImage = styled.div`
-  position: absolute;
-  height: 100%;
-  width: 50%;
-  top: 0;
-  right: 0;
-  z-index: -1;
-  border-top: solid 200px transparent;
-  background-image: url(${props => props.src});
-  background-size: contain;
-  background-repeat: no-repeat;
-  background-position: right top;
-
-  @media screen and (max-width: 768px) {
-    display: none;
-  }
-`;
-
-const TextContainer = styled.div`
-  max-width: 800px;
-  padding: 64px 112px;
-  width: 70%;
-  font-size: 14px;
-
-  h2 {
-    font: bold 32px/48px;
-    margin: 24px 0 16px;
-    position: relative;
-  }
-  h3 {
-    font: bold 16px/24px;
-    margin: 16px 0 0;
-    position: relative;
-  }
-  h3 > img {
-    position: absolute;
-    top: -4px;
-    width: 36px;
-    left: -48px;
-  }
-  hr {
-    border: none;
-    background: #e1e8f0;
-    height: 1px;
-    margin: 24px 0 0;
-    width: 32px;
-    height: 2px;
-  }
-  @media screen and (max-width: 768px) {
-    max-width: 100%;
-    width: 100%;
-    padding: 48px 48px 48px 80px;
-  }
-`;
 
 export default function IndexPage() {
   const baseUrl = useBaseUrl('/');
@@ -83,8 +26,11 @@ export default function IndexPage() {
     >
       <Home HeroExample={HeroExample}>
         <div style={{position: 'relative'}}>
-          <FeatureImage src={`${baseUrl}images/maps.jpg`} />
-          <TextContainer>
+          <div
+            className={styles.featureImage}
+            style={{backgroundImage: `url(${baseUrl}images/maps.jpg)`}}
+          />
+          <div className={styles.textContainer}>
             <h2>
               deck.gl-community is a set of experimental add-on modules for deck.gl.
             </h2>
@@ -112,7 +58,7 @@ export default function IndexPage() {
 
             </p>
 
-          </TextContainer>
+          </div>
         </div>
       </Home>
     </Layout>